+++ test/parser_test.rb
@@ -1,16 +1,16 @@
 require 'test_helper'
 
-class CfdiParserTest < MiniTest::Test
+class ParserTest < MiniTest::Test
 
   def test_is_version_3_2?
     file = File.read(File.expand_path('../fixtures/cfdi_3_2.xml', __FILE__))
-    @cfdi_parser = CfdiParser::CfdiParser.new(file)
+    @cfdi_parser = CfdiParser::Parser.new(file)
     assert(@cfdi_parser.is_version_3_2?, 'must be true')
   end
 
   def test_is_version_3_3?
     file = File.read(File.expand_path('../fixtures/cfdi_3_3.xml', __FILE__))
-    @cfdi_parser = CfdiParser::CfdiParser.new(file)
+    @cfdi_parser = CfdiParser::Parser.new(file)
     assert(@cfdi_parser.is_version_3_3?, 'must be true')
   end
 end
